+++ src/server/common/components/marine-licence/activity-details-card/template.njk
@@ -39,7 +39,7 @@
 {% set _ = rows.push({
     key: { text: "Maximum duration of activity"},
     value: { html: appIncompleteField({text:params.activityDetails.activityDuration})},
-    actions: setSiteDetailsAction(params.activityDetails.activityDuration, 'maximum-duration', params.siteNumber, "", { skipAction: true, activityNumber: params.index })
+    actions: setSiteDetailsAction(params.activityDetails.activityDuration, 'activity-duration', params.siteNumber, "", { skipAction: true, activityNumber: params.index })
 }) %}
 
 {% set _ = rows.push({
