+++ src/server/marine-licence/site-details/activity-duration/controller.js
@@ -4,7 +4,7 @@ import {
 } from '#src/server/common/helpers/marine-licence/session-cache/utils.js'
 import { marineLicenceRoutes } from '#src/server/common/constants/routes.js'
 import { getActivityDetailsByIndex } from '#src/server/common/helpers/marine-licence/session-cache/site-details-utils.js'
-import { durationSchema } from '#src/server/marine-licence/site-details/maximum-duration/schema.js'
+import { durationSchema } from '#src/server/marine-licence/site-details/activity-duration/schema.js'
 import { getSiteDataFromParam } from '#src/server/common/helpers/site-details/site-name.js'
 import { createFailAction } from '#src/server/common/helpers/createFailAction.js'
 
@@ -13,7 +13,7 @@ export const durationErrorMessages = {
 }
 
 export const MARINE_LICENCE_DURATION_VIEW_ROUTE =
-  'marine-licence/site-details/maximum-duration/index'
+  'marine-licence/site-details/activity-duration/index'
 
 const getBackLink = (siteNumber, activityDetailsNumber) =>
   `${marineLicenceRoutes.MARINE_LICENCE_REVIEW_SITE_DETAILS}#activity-details-site-${siteNumber}-activity-${activityDetailsNumber}`
@@ -23,7 +23,7 @@ export const durationSettings = {
   heading: 'What is the maximum duration of the activity?'
 }
 
-export const durationController = {
+export const activityDurationController = {
   handler(request, h) {
     const marineLicence = getMarineLicenceCache(request)
 
@@ -54,7 +54,7 @@ export const durationController = {
   }
 }
 
-export const durationSubmitController = {
+export const activityDurationSubmitController = {
   options: {
     validate: {
       payload: durationSchema,
