+++ src/server/marine-licence/site-details/activity-duration/index.js
@@ -1,18 +1,18 @@
 import {
-  durationController,
-  durationSubmitController
-} from '#src/server/marine-licence/site-details/maximum-duration/controller.js'
+  activityDurationController,
+  activityDurationSubmitController
+} from '#src/server/marine-licence/site-details/activity-duration/controller.js'
 import { marineLicenceRoutes } from '#src/server/common/constants/routes.js'
 
 export const durationRoutes = [
   {
     method: 'GET',
     path: marineLicenceRoutes.MARINE_LICENCE_DURATION,
-    ...durationController
+    ...activityDurationController
   },
   {
     method: 'POST',
     path: marineLicenceRoutes.MARINE_LICENCE_DURATION,
-    ...durationSubmitController
+    ...activityDurationSubmitController
   }
 ]
